01-tutorials/01-fundamentals/09-simple-observability/requirements.txt
strands-agents
strands-agents-tools
ddgs
arize-otel
arize-toolkit
strands-agents[openai]
arize-otel 
openinference-instrumentation-openai 
opentelemetry-exporter-otlp